Add header menu and refactor ui to us app.
