+++ app-us/package.json
@@ -39,6 +39,7 @@
         "@textea/json-viewer": "^2.14.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
+        "react-icons": "^5.3.0",
         "react-router-dom": "^6.3.0"
     },
     "devDependencies": {
